Moved custom template functions to presenters
