+++ presenters/dashboard.go
@@ -1,0 +1,19 @@
+package presenters
+
+import "google-scraper/models"
+
+const (
+	layoutUS = "January 2, 2006"
+)
+
+func DisplayStatus(k models.Keyword) string {
+	if k.SearchCompleted {
+		return "Yes"
+	} else {
+		return "No"
+	}
+}
+
+func DisplayFormattedCreatedDate(k models.Keyword) string {
+	return k.CreatedAt.Format(layoutUS)
+}
